Extract tracker card color logic into testable hook
Break TrackerCardColorSettings into a pure manager module
(tracker-card-color-manager.ts), a React state hook
(use-tracker-card-color-manager.ts), and unit tests covering
cache patching and target resolution. The component is now a
thin render-only shell. No behavior changes.
